Update curve_item.py
Added a number of small modifications to make the 'correction factors' feature work in a useful way. Unfortunately, we will use a bit more of memory, but the feature becomes useful. Before it would only apply the correction factors to the newly acquired data, forcing you to retest an axis from end to end again. Now the update is done on all the capture data.

Besides support for local minimum and maximum (between haircrosses) has been added.